structure folder to use maven for compiling and running with jdbc
